# docs/api/openapi.json
{
 "components": {
  "schemas": {
   "AskRequest": {
    "additionalProperties": false,
    "description": "One question, about ONE document version.\n\nLength-bounded at the boundary; content rules live server-side.\n\n``document_version_id`` names the version the question is about \u2014 the one the\nreader has open. It is OPTIONAL and defaults to the contract's newest version,\nwhich is what every caller got before this field existed. Authorization is not\nweakened by naming it: the version is resolved through `guard.document_version`\nand additionally required to belong to the conversation's own contract, so this\nfield can only ever narrow the scope, never widen it.",
    "properties": {
     "document_version_id": {
      "anyOf": [
       {
        "maxLength": 64,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Document Version Id"
     },
     "question": {
      "maxLength": 2000,
      "minLength": 1,
      "title": "Question",
      "type": "string"
     }
    },
    "required": [
     "question"
    ],
    "title": "AskRequest",
    "type": "object"
   },
...
   "ContractUpdate": {
    "additionalProperties": false,
    "properties": {
     "contract_type": {
      "anyOf": [
       {
        "maxLength": 200,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contract Type"
     },
     "name": {
      "anyOf": [
       {
        "maxLength": 500,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "status": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ContractStatus"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "title": "ContractUpdate",
    "type": "object"
   },
   "ConversationCreate": {
    "additionalProperties": false,
    "description": "An assist-lane session, optionally scoped to a contract the requester can view.",
    "properties": {
     "contract_id": {
      "anyOf": [
       {
        "maxLength": 64,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contract Id"
     }
    },
    "title": "ConversationCreate",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "LegalMind API",
  "version": "1"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/contracts/{contract_id}": {
   "get": {
    "description": "The contract with its document versions, newest first.\n\n`document_versions` is an implementation addition (2026-08-30, UI phase): a\ndocument-anchored workspace opened on a contract must be able to find its\ndocument through the API, and nothing listed versions \u2014 the legacy screen only\never showed the version it had just uploaded. Additive, same permission the\ncontract already required (a version is reachable only through a contract the\ncaller can see \u2014 47.6 one level down), and the summary is the existing\n`serialize_document_version` shape, so nothing new is disclosed. Recorded in\nStep 49's implementation-additions section.",
    "operationId": "get_contract_api_v1_contracts__contract_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "contract_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Contract Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Contract Api V1 Contracts  Contract Id  Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Contract",
    "tags": [
     "contracts"
    ]
   },
   "patch": {
    "operationId": "update_contract_api_v1_contracts__contract_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "contract_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Contract Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ContractUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Update Contract Api V1 Contracts  Contract Id  Patch",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Update Contract",
    "tags": [
     "contracts"
    ]
   }
  },
...
  "/api/v1/conversations": {
   "get": {
    "description": "The caller's own conversations, newest first \u2014 49.6 r4 applied to the assist\nlane: the list carries exactly what `GET /conversations/{id}` would return and\nnothing a GET would 404 on, so it cannot become the enumeration oracle `AM-25`\nr7 forbids. `contract_id` is the one allow-listed filter (49.6 r3): the workspace\nfor a document shows that document's history and no other.",
    "operationId": "list_conversations_api_v1_conversations_get",
    "parameters": [
     {
      "in": "query",
      "name": "contract_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "uuid",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Contract Id"
      }
     },
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 25,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response List Conversations Api V1 Conversations Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Conversations",
    "tags": [
     "assist"
    ]
   },
   "post": {
    "operationId": "create_conversation_api_v1_conversations_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ConversationCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Create Conversation Api V1 Conversations Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Conversation",
    "tags": [
     "assist"
    ]
   }
  },
  "/api/v1/conversations/{conversation_id}": {
   "get": {
    "description": "A conversation with its turns \u2014 and, for every assistant turn, the SAME\ncitation objects the live answer carried, so a reload renders exactly what was\nshown (`AM-25` r5: every answer traceable to its retrieved evidence, on every\nview of it, not only the first).\n\nCitations are rebuilt from `answer_citations` (the verified claim\u2192chunk rows),\nthe chunk's evidence row (page, section) and the retrieval run (the score for\nthat chunk in THAT query \u2014 a property of the query, so it lives on the run, not\nthe chunk). Refusals and routed turns have no citation rows and get `[]`.",
    "operationId": "get_conversation_api_v1_conversations__conversation_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "conversation_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Conversation Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Conversation Api V1 Conversations  Conversation Id  Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Conversation",
    "tags": [
     "assist"
    ]
   }
  },
  "/api/v1/conversations/{conversation_id}/messages": {
   "post": {
    "operationId": "ask_api_v1_conversations__conversation_id__messages_post",
    "parameters": [
     {
      "in": "path",
      "name": "conversation_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Conversation Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AskRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Ask Api V1 Conversations  Conversation Id  Messages Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Ask",
    "tags": [
     "assist"
    ]
   }
  },
...
  "/api/v1/document-versions/{document_version_id}/extract-obligations": {
   "post": {
    "description": "Key Obligations extraction (owner, 2026-08-31) \u2014 descriptive facts about\nthe document's own text, never a Finding or a judgment. Runs synchronously\n(the Ask precedent: one bounded generation call); idempotent-by-refusal \u2014\na version already extracted keeps its record, since the version is\nimmutable and re-running buys nothing.\n\nPermission is `finding.view`: obligations are facts about text the caller\ncan already read in full, not the organization's negotiating position, so\nLEGAL-02's stricter gate does not apply.",
    "operationId": "extract_obligations_api_v1_document_versions__document_version_id__extract_obligations_post",
    "parameters": [
     {
      "in": "path",
      "name": "document_version_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Document Version Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Extract Obligations Api V1 Document Versions  Document Version Id  Extract Obligations Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Extract Obligations",
    "tags": [
     "assist"
    ]
   }
  },
  "/api/v1/document-versions/{document_version_id}/obligations": {
   "get": {
    "operationId": "read_obligations_api_v1_document_versions__document_version_id__obligations_get",
    "parameters": [
     {
      "in": "path",
      "name": "document_version_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Document Version Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Read Obligations Api V1 Document Versions  Document Version Id  Obligations Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Read Obligations",
    "tags": [
     "assist"
    ]
   }
  },
  "/api/v1/document-versions/{document_version_id}/suggest-type": {
   "post": {
    "description": "Assist-lane type suggestion (owner, 2026-08-31) \u2014 a proposal, never a write.\n\nThe Guard resolves the version through its contract (47.6), the assist lane\nreads the opening evidence, and the response is only ever advice for the UI's\npre-fill. The type itself is recorded exclusively by the human's own\n`PATCH /contracts/{id}` \u2014 Q9's substance (human-declared) survives intact.\nEvery failure shape is the same honest `confident: false`, so the screen\ndegrades to the plain empty select rather than erroring.",
    "operationId": "suggest_type_api_v1_document_versions__document_version_id__suggest_type_post",
    "parameters": [
     {
      "in": "path",
      "name": "document_version_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Document Version Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Suggest Type Api V1 Document Versions  Document Version Id  Suggest Type Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Suggest Type",
    "tags": [
     "assist"
    ]
   }
  },
...
  "/api/v1/reviews/{review_id}": {
   "get": {
    "operationId": "get_review_api_v1_reviews__review_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "review_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Review Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Review Api V1 Reviews  Review Id  Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Review",
    "tags": [
     "reviews"
    ]
   }
  },
...
 }
}